refactor: move file into pkg folder for shared package
