+++ pkg/triton/const.go
