+++ pkg/triton/triton.go
@@ -19,7 +19,7 @@ import (
 	"google.golang.org/grpc/credentials/insecure"
 
 	"github.com/instill-ai/model-backend/config"
-	"github.com/instill-ai/model-backend/internal/inferenceserver"
+	"github.com/instill-ai/model-backend/pkg/triton/inferenceserver"
 
 	modelPB "github.com/instill-ai/protogen-go/vdp/model/v1alpha"
 )
